III-3949 - Rename EventForm to OfferForm
